requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
boto3==1.26.77
botocore==1.29.77
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-cors-headers==3.13.0
django-db-logger==0.1.12
django-excel==0.0.10
django-import-export==3.0.2
django-multiselectfield==0.1.12
django-simple-history==3.2.0
django-storages==1.13.2
django-webpush==0.3.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
html5lib==1.1
http-ece==1.1.0
idna==3.4
jmespath==1.0.1
lml==0.1.0
lxml==4.9.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
oscrypto==1.3.0
passlib==1.7.4
Pillow==9.4.0
psycopg2-binary==2.9.5
py-vapid==1.9.0
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyHanko==0.17.0
pyhanko-certvalidator==0.20.0
PyJWT==2.6.0
pypdf==3.4.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywebpush==1.9.4
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tablib==3.3.0
texttable==1.6.7
tinycss2==1.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
webencodings==0.5.1
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
pandas
django-filter==2.4.0
tabulate
beautifulsoup4==4.12.2
razorpay==1.3.0
PyVimeo==1.1.0
channels==4.0.0
daphne==4.0.0
redis==4.5.5
channels-redis==4.1.0
djangochannelsrestframework
gunicorn==20.1.0
celery==5.3.1
django-redis==5.3.0
sentry-sdk==1.29.2
django_celery_beat